ai_job_screening/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
python-docx==1.0.0
PyPDF2==3.0.1
